# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Oxidean REST API
  version: '1'
  description: Public REST API for Oxidean — a companion to the typed JSON RPC at `POST /api/rpc`. Each route dispatches to the same internal procedure the RPC layer uses, so ACLs, PAT scope gates, and side effects are identical. Generated from the route table — this document cannot drift from the served surface.
servers:
- url: /api/v1
security:
- sessionCookie: []
- bearerPat: []
...
paths:
...
  /repos/{owner}/{repo}/pulls:
    get:
      tags:
      - pulls
      summary: List pull requests (`pull.list`)
      operationId: listPulls
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: state
        in: query
        required: false
        schema:
          description: '`open` (default) | `closed` | `merged` | `all`'
          type:
          - string
          - 'null'
          default: null
      - name: author
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
          default: null
      - name: label
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
          default: null
      - name: assignee
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
          default: null
      - name: q
        in: query
        required: false
        schema:
          description: Title/body substring search (same semantics as issue.list `q`).
          type:
          - string
          - 'null'
          default: null
      - name: review_state
        in: query
        required: false
        schema:
          description: '`approved` | `changes_requested` | `review_required` | …'
          type:
          - string
          - 'null'
          default: null
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: uint32
          minimum: 0
          default: null
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: uint32
          minimum: 0
          default: null
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullListResponse'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - pulls
      summary: Create pull request (`pull.create`)
      operationId: createPull
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
              - title
              - base_ref
              - head_ref
              title: CreatePullRequest
              type: object
              properties:
                draft:
                  type:
                  - boolean
                  - 'null'
                  default: null
                head_name:
                  type:
                  - string
                  - 'null'
                  default: null
                title:
                  type: string
                body:
                  type:
                  - string
                  - 'null'
                  default: null
                base_ref:
                  type: string
                head_ref:
                  type: string
                head_owner:
                  description: Optional fork head owner (defaults to base owner / same-repo).
                  type:
                  - string
                  - 'null'
                  default: null
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullPublic'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /repos/{owner}/{repo}/pulls/{number}:
    get:
      tags:
      - pulls
      summary: Get pull request (`pull.get`)
      operationId: getPull
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullPublic'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
      - pulls
      summary: Update pull (`pull.update` + `pull.close`/`pull.reopen` when `state` is given)
      operationId: updatePull
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
                base_ref:
                  type: string
                draft:
                  type: boolean
                state:
                  type: string
                  enum:
                  - open
                  - closed
                  description: '`closed` → pull.close; `open` → pull.reopen'
              description: At least one of `title`, `body`, `base_ref`, `draft`, `state` is required
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullPublic'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /repos/{owner}/{repo}/pulls/{number}/merge:
    post:
      tags:
      - pulls
      summary: Merge pull request (`pull.merge`)
      operationId: mergePull
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
              - method
              title: MergePullRequest
              type: object
              properties:
                delete_branch:
                  type:
                  - boolean
                  - 'null'
                  default: null
                commit_message:
                  type:
                  - string
                  - 'null'
                  default: null
                commit_title:
                  type:
                  - string
                  - 'null'
                  default: null
                method:
                  $ref: '#/components/schemas/MergeMethod'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MergePullResponse'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /repos/{owner}/{repo}/pulls/{number}/files:
    get:
      tags:
      - pulls
      summary: Pull diff files (`pull.files`)
      operationId: listPullFiles
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullFilesResponse'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /repos/{owner}/{repo}/pulls/{number}/commits:
    get:
      tags:
      - pulls
      summary: Pull commits (`pull.commits`)
      operationId: listPullCommits
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullCommitsResponse'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /repos/{owner}/{repo}/pulls/{number}/comments:
    get:
      tags:
      - pulls
      summary: Pull review comments (`pull.comments.list`)
      operationId: listPullComments
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullCommentsListResponse'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - pulls
      summary: Add pull comment (`pull.comments.create`; diff placement optional)
      operationId: createPullComment
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
              - body
              title: CreatePullCommentRequest
              type: object
              properties:
                commit_sha:
                  type:
                  - string
                  - 'null'
                  default: null
                start_line:
                  type:
                  - integer
                  - 'null'
                  format: int64
                  default: null
                line:
                  type:
                  - integer
                  - 'null'
                  format: int64
                  default: null
                body:
                  type: string
                path:
                  type:
                  - string
                  - 'null'
                  default: null
                side:
                  type:
                  - string
                  - 'null'
                  default: null
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullCommentPublic'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /repos/{owner}/{repo}/pulls/{number}/reviews:
    get:
      tags:
      - pulls
      summary: Pull reviews (`pull.reviews.list`)
      operationId: listPullReviews
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullReviewsListResponse'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - pulls
      summary: Submit review (`pull.reviews.submit`)
      operationId: submitPullReview
      parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
              - state
              title: SubmitPullReviewRequest
              type: object
              properties:
                body:
                  type:
                  - string
                  - 'null'
                  default: null
                state:
                  description: '`approved` | `changes_requested` | `commented`'
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullReviewPublic'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: domain.error_code
        message:
          type: string
        data:
          description: optional structured detail
...
    MergeMethod:
      type: string
      enum:
      - merge
      - squash
      - rebase
...
    PullListResponse:
      type: object
      properties:
        pulls:
          type: array
          items:
            $ref: '#/components/schemas/PullPublic'
        total:
          type: integer
          format: int64
        offset:
          type: integer
          format: uint32
          minimum: 0
        limit:
          type: integer
          format: uint32
          minimum: 0
      required:
      - pulls
      - total
      - offset
      - limit
    PullPublic:
      type: object
      properties:
        id:
          type: string
        repo_id:
          type: string
        number:
          type: integer
          format: int64
        title:
          type: string
        body:
          type: string
        state:
          $ref: '#/components/schemas/PullState'
        draft:
          type: boolean
        author_id:
          type: string
        author_username:
          type: string
        base_ref:
          type: string
        base_sha:
          type: string
        head_repo_id:
          type: string
        head_owner:
          type: string
        head_name:
          type: string
        head_ref:
          type: string
        head_sha:
          type: string
        merged_at:
          type:
          - string
          - 'null'
        merged_by:
          type:
          - string
          - 'null'
        merge_commit_sha:
          type:
          - string
          - 'null'
        merge_method:
          anyOf:
          - $ref: '#/components/schemas/MergeMethod'
          - type: 'null'
        closed_at:
          type:
          - string
          - 'null'
        closed_by:
          type:
          - string
          - 'null'
        created_at:
          type: string
        updated_at:
          type: string
        comment_count:
          type: integer
          format: int64
          default: 0
        assignees:
          type: array
          items:
            $ref: '#/components/schemas/IssueAssigneePublic'
          default: []
      required:
      - id
      - repo_id
      - number
      - title
      - body
      - state
      - draft
      - author_id
      - author_username
      - base_ref
      - base_sha
      - head_repo_id
      - head_owner
      - head_name
      - head_ref
      - head_sha
      - created_at
      - updated_at
...
    MergePullResponse:
      type: object
      properties:
        pull:
          $ref: '#/components/schemas/PullPublic'
        merge_commit_sha:
          type: string
      required:
      - pull
      - merge_commit_sha
    PullFilesResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/PullDiffFile'
        empty:
          type: boolean
      required:
      - files
      - empty
...
    PullCommitsResponse:
      type: object
      properties:
        commits:
          type: array
          items:
            $ref: '#/components/schemas/PullCommitSummary'
      required:
      - commits
...
    PullCommentsListResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/PullCommentPublic'
      required:
      - comments
    PullCommentPublic:
      type: object
      properties:
        id:
          type: string
        pull_id:
          type: string
        author_id:
          type: string
        author_username:
          type: string
        body:
          type: string
        path:
          type:
          - string
          - 'null'
        side:
          type:
          - string
          - 'null'
        line:
          type:
          - integer
          - 'null'
          format: int64
        start_line:
          type:
          - integer
          - 'null'
          format: int64
        commit_sha:
          type:
          - string
          - 'null'
        outdated:
          type: boolean
        resolved:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
      - id
      - pull_id
      - author_id
      - author_username
      - body
      - outdated
      - resolved
      - created_at
      - updated_at
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oxidean_session
      description: Browser session cookie — wins if both credentials are sent.
    bearerPat:
      type: http
      scheme: bearer
      description: Personal access token. Classic `repo` covers the repo domain; `admin.*` and org/credential mutations are session-only.